+++ MonoRemoteDebugger.SharedLib/Server/MonoDesktopProcess.cs
@@ -19,7 +19,7 @@ namespace MonoRemoteDebugger.SharedLib.Server
 
             string args = GetProcessArgs();
             ProcessStartInfo procInfo = GetProcessStartInfo(workingDirectory, monoBin);
-            procInfo.Arguments = args + string.Format(" --config \"{0}.config\" \"{0}\"", _targetExe);
+            procInfo.Arguments = args + string.Format(" --config \"{0}.config\" \"{0}\" {1}", _targetExe, Arguments);
 
             _proc = Process.Start(procInfo);
             RaiseProcessStarted();